requirements.txt
google-auth==2.35.0
requests>=2.32.0
psycopg[binary]==3.2.3
psycopg_pool==3.2.4
python-dotenv==1.0.0
